refactor(research): extract EPUB archive packaging
